Use config.ready() to ensure initial response
Previously, if subscribe was called prior to onUpdate resolving, there would be no initial location to pass as an argument to the subscribed function. That defeated the purpose of that function, so now the ready function can be used to ensure that a response exists. It is possible that this will need to be tweaked (possibly to deal with route refreshes), but it works for now.