Addons receive internal route object
The internal route object is similar to the one provided by the user, but has a render function instead of a value/call property. It also has a match function and a keys array. This change allows addons to access these values. The only breaking change is that you can no longer access the route's value/call directly. I don't think that this is an issue, but it can be remedied later if necessary.