load(params, responseCreator)
Changed the API so that it makes more sense to call load with the prefetch addon. Instead of mocking a response creator, only the params are ncessary. If the load function is called by the prefetch addon but needs the responseCreator, it should know to check if a responseCreator is defined.