Fix parent matches when no children match
The old behavior was that the route would still match even if none of its children did. That was essentially forcing non-end matching on the route if it has children. Now, if no children match, then the parent route will only match if it is an exact match or the user has explicitly stated that the route is a non-end route.