Remove external-helpers deps
Eventually I would like to add this back in, but I'll need to figure out how to get the dependency injection working as desired. There also seems to be some issues with the UMD builds because external-helpers was not making a difference with the curi-react UMD build. Not a priority, but getting rid of the warnings during builds would be nice.