Make curi-addon-active a dependency of curi-react-active
This is mostly because I was tired of the NPM warnings when curi-addon-active is not installed. This technically isn't a dependency because you will get a warning when attempting to use the <Active> component without installing the active addon, but this should simplify things since the user can expect the addon to be installed for them.