(Docs) Moving the site back to gh-pages
I am copying the format that Facebook takes here. The code for generating the documentation website exists in the master branch. The generated HTML (and static files) will exist in the gh-pages branch. In order to add the files to the gh-pages branch, I will just keep a clone of the Curi repo and copy the files over to that directory. Not the prettiest solution, but it should work well enough and keep documentation builds out of the main commit history.