Switch to <Navigator render={fn} />
This enforces one way to render whereas with children there were two ways to do this. There wasn't a real benefit to having multiple ways to render the children, so it should be easier to have one approach that is used in all documentation. Also, it is a 'render function', so it makes sense to call the prop 'render'.