Add route.match API
The "load", "preload", and "body" properties of a route have been removed. Instead, there is a "match" object, which has "initial", "every", and "finish" properties.

"initial" is equivalent to the old "preload", while the old "load" is now split between the "every" and "finish" functions.

The old "body" function is now just a function of the "modifiers" object passed to "finish" functions. Since "finish" has access to what the "initial" function resolved, we can just "modifiers.body(resolved.initial) to attach a body property to a response. This gets rid of the need to keep a component store.