Use react-broadcast to mimic new context API (#35)
Take advantage of the fact that this is still pre v1 final to switch to using the new React context API. This isn't _really_ using the new context, but puts the pieces in place to easily do that once it is official. This also cleans up the API.

The fate of the @curi/redux and @curi/mobx packages are up in the air. They won't be deprecated yet, but with the new context API, they are less useful.